requirements.txt
fastapi
uvicorn
passlib
mysql-connector-python
bcrypt 
python-jose[cryptography] 
passlib